+++ src/elements/db/ElementQuery.php
@@ -166,10 +166,10 @@ class ElementQuery extends Query implements ElementQueryInterface
     public bool $savedDraftsOnly = false;
 
     /**
-     * @var bool Whether revision elements should be returned.
+     * @var bool|null Whether revision elements should be returned.
      * @since 3.2.0
      */
-    public bool $revisions = false;
+    public ?bool $revisions = false;
 
     /**
      * @var int|null The ID of the revision to return (from the `revisions` table)
@@ -727,7 +727,7 @@ class ElementQuery extends Query implements ElementQueryInterface
      * @inheritdoc
      * @uses $revisions
      */
-    public function revisions(bool $value = true): self
+    public function revisions(?bool $value = true): self
     {
         $this->revisions = $value;
         return $this;
@@ -741,7 +741,9 @@ class ElementQuery extends Query implements ElementQueryInterface
     public function revisionId(?int $value = null): self
     {
         $this->revisionId = $value;
-        $this->revisions = $value !== null;
+        if ($value !== null && $this->revisions === false) {
+            $this->revisions = true;
+        }
         return $this;
     }
 
@@ -759,7 +761,9 @@ class ElementQuery extends Query implements ElementQueryInterface
         } else {
             throw new InvalidArgumentException('Invalid revisionOf value');
         }
-        $this->revisions = $value !== null;
+        if ($value !== null && $this->revisions === false) {
+            $this->revisions = true;
+        }
         return $this;
     }
 
@@ -777,7 +781,9 @@ class ElementQuery extends Query implements ElementQueryInterface
         } else {
             throw new InvalidArgumentException('Invalid revisionCreator value');
         }
-        $this->revisions = $value !== null;
+        if ($value !== null && $this->revisions === false) {
+            $this->revisions = true;
+        }
         return $this;
     }
 
@@ -1814,12 +1820,20 @@ class ElementQuery extends Query implements ElementQueryInterface
             }
         }
 
-        if ($this->revisions) {
-            $behaviors['revision'] = new RevisionBehavior([
-                'creatorId' => ArrayHelper::remove($row, 'revisionCreatorId'),
-                'revisionNum' => ArrayHelper::remove($row, 'revisionNum'),
-                'revisionNotes' => ArrayHelper::remove($row, 'revisionNotes'),
-            ]);
+        if ($this->revisions !== false) {
+            if (!empty($row['revisionId'])) {
+                $behaviors['revision'] = new RevisionBehavior([
+                    'creatorId' => ArrayHelper::remove($row, 'revisionCreatorId'),
+                    'revisionNum' => ArrayHelper::remove($row, 'revisionNum'),
+                    'revisionNotes' => ArrayHelper::remove($row, 'revisionNotes'),
+                ]);
+            } else {
+                unset(
+                    $row['revisionCreatorId'],
+                    $row['revisionNum'],
+                    $row['revisionNotes'],
+                );
+            }
         }
 
         $element = new $class($row);
@@ -2467,16 +2481,17 @@ class ElementQuery extends Query implements ElementQueryInterface
             $this->subQuery->andWhere($this->_placeholderCondition(['elements.draftId' => null]));
         }
 
-        if ($this->revisions) {
-            $this->subQuery->innerJoin(['revisions' => Table::REVISIONS], '[[revisions.id]] = [[elements.revisionId]]');
-            $this->query
-                ->innerJoin(['revisions' => Table::REVISIONS], '[[revisions.id]] = [[elements.revisionId]]')
-                ->addSelect([
-                    'elements.revisionId',
-                    'revisions.creatorId as revisionCreatorId',
-                    'revisions.num as revisionNum',
-                    'revisions.notes as revisionNotes',
-                ]);
+        if ($this->revisions !== false) {
+            $joinType = $this->revisions === true ? 'INNER JOIN' : 'LEFT JOIN';
+            $this->subQuery->join($joinType, ['revisions' => Table::REVISIONS], '[[revisions.id]] = [[elements.revisionId]]');
+            $this->query->join($joinType, ['revisions' => Table::REVISIONS], '[[revisions.id]] = [[elements.revisionId]]');
+
+            $this->query->addSelect([
+                'elements.revisionId',
+                'revisions.creatorId as revisionCreatorId',
+                'revisions.num as revisionNum',
+                'revisions.notes as revisionNotes',
+            ]);
 
             if ($this->revisionId) {
                 $this->subQuery->andWhere(['elements.revisionId' => $this->revisionId]);
@@ -2801,8 +2816,7 @@ class ElementQuery extends Query implements ElementQueryInterface
         }
         $caseSql .= ' else ' . count($preferSites) . ' end';
 
-        $subSelectSqlQuery = clone $this->subQuery;
-        $subSelectSql = $subSelectSqlQuery
+        $subSelectSql = (clone $this->subQuery)
             ->select(['elements_sites.id'])
             ->andWhere('[[subElements.id]] = [[tmpElements.id]]')
             ->orderBy([